ami_coding_pari_na/requirements.txt
asgiref==3.6.0
Django==4.1.4
Pillow==9.3.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
djangorestframework==3.14.0
requests==2.28.1
